backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
shap
pydantic
requests
beautifulsoup4
apscheduler
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
